s5_continuous_integration/small_project/requirements.txt
# local package
-e .
torch>=1.8
torchvision
click
matplotlib
black
isort
pytest

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
